Update Security For Api

+++ src/main/java/com/springbook/booksale/entity/TopicEntity.java
@@ -6,6 +6,10 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.Set;
 
 @Entity
 @Getter
@@ -17,7 +21,7 @@ public class TopicEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "id")
-    private int id;
+    private long id;
     @Basic
     @Column(name = "name")
     private String name;
@@ -25,11 +29,11 @@ public class TopicEntity {
     @Column(name = "is_del", columnDefinition = "bit default false")
     private boolean isDel;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
